Fix page overlay bug
This fixes the page overlay bug #23 I guess.
The changes to composer.json can be ignored. I had to make them to
temporarily integrate my fork into our project.

The problem with the page overlay was, that $this->cObj->data only
contains the information of the current database record, but without
being overlayed. The $GLOBALS['TSFE']->page is the way to go for all
fields.

Additionally fetching the file relations have to be done from
pages_language_overlay if not in default language. Therefore I
introduced two new variables for fetching file relations based on the
curreent sys_language_uid.

This is only tested on TYPO3 7.6.9 but I think it should work on every
other supported version.